Move the convert tests into a separate file.
